requirements.txt
Flask==1.0.2
Werkzeug<0.13.0,>=0.12.0
google-cloud-bigquery==0.31.0
google-api-core==1.1.0
web3==4.0.0